Build with Rust 1.0.0-alpha
Since the `impl_keyed_iter` was not exported and it was incomplete
(i.e.: it assumed idents instead of ty for K:), I dropped it in
favor of inlining the impl.

+++ tests/groupable-tests.rs
@@ -1,4 +1,3 @@
-#![feature(macro_rules)]
 extern crate groupable;
 
 use std::collections::{HashMap, VecMap};
@@ -34,8 +33,8 @@ macro_rules! test_char_vec (
 
 #[test]
 fn hashmap() {
-    test_string!(HashMap<uint, String>);
-    test_char_vec!(HashMap<uint, Vec<char>>);
+    test_string!(HashMap<usize, String>);
+    test_char_vec!(HashMap<usize, Vec<char>>);
 }
 
 #[test]
